Good enough version
So it does mimick the clock from NGE, the menu does cronometer, countdown, it resets the timer and the special Racing does cool stuff. And to go back to the clock you click on itself.

I want to make it more useful as an start page with RSS reader and weather stuff.